# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Flyff Idle API
  version: 0.1.0
  description: API schema for auth, characters, inventory, admin actions, and game data.
servers:
  - url: http://localhost:4000
...
paths:
...
  /api/characters/{characterId}/progression/stat-allocations:
    post:
      tags:
        - Characters
      summary: Allocate available character stat points
      operationId: allocateCharacterStats
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatAllocationRequest"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/DomainRuleFailed"
  /api/characters/{characterId}/progression/skill-allocations:
    post:
      tags:
        - Characters
      summary: Allocate available character skill points
      operationId: allocateCharacterSkills
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillAllocationRequest"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/DomainRuleFailed"
  /api/characters/{characterId}/progression/battle-state:
    put:
      tags:
        - Characters
      summary: Replace the persisted progression state from the client-simulated battle system
      operationId: replaceCharacterBattleState
      description: This is an idempotent persistence boundary for the current client-simulated battle implementation. It is not an authoritative battle-resolution endpoint.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BattleProgressionStateRequest"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/characters/{characterId}/inventory/move:
    post:
      tags:
        - Characters
      summary: Move, swap, or stack inventory items
      operationId: moveCharacterInventoryItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveInventoryItemRequest"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/characters/{characterId}/inventory/sort:
    post:
      tags:
        - Characters
      summary: Sort a character inventory
      operationId: sortCharacterInventory
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SortInventoryRequest"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/characters/{characterId}/inventory/{slotIndex}/consume:
    post:
      tags:
        - Characters
      summary: Consume an item directly from an inventory slot
      operationId: consumeCharacterInventoryItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
        - $ref: "#/components/parameters/SlotIndexPath"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/characters/{characterId}/inventory/loot:
    post:
      tags:
        - Characters
      summary: Add awarded loot to a character inventory
      operationId: addCharacterLoot
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LootInventoryRequest"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    CharacterIdPath:
      in: path
      name: characterId
      required: true
      schema:
        type: string
    SlotIndexPath:
      in: path
      name: slotIndex
      required: true
      schema:
        type: integer
        minimum: 0
        maximum: 99
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid bearer token, or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    DomainRuleFailed:
      description: The request was valid but violates a domain rule
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CharacterResponse:
      type: object
      required:
        - character
      properties:
        character:
          $ref: "#/components/schemas/Character"
...
    StatAllocationRequest:
      type: object
      required:
        - allocations
      properties:
        allocations:
          type: object
          required:
            - str
            - sta
            - dex
            - int
          properties:
            str:
              type: integer
              minimum: 0
              maximum: 984
            sta:
              type: integer
              minimum: 0
              maximum: 984
            dex:
              type: integer
              minimum: 0
              maximum: 984
            int:
              type: integer
              minimum: 0
              maximum: 984
    SkillAllocationRequest:
      type: object
      required:
        - allocations
      properties:
        allocations:
          type: object
          minProperties: 1
          additionalProperties:
            type: integer
            minimum: 1
            maximum: 20
    BattleProgressionStateRequest:
      type: object
      required:
        - exp
        - level
        - penya
      properties:
        exp:
          type: integer
          minimum: 0
        level:
          type: integer
          minimum: 1
          maximum: 170
        penya:
          type: integer
          minimum: 0
...
    MoveInventoryItemRequest:
      type: object
      required:
        - fromSlotIndex
        - toSlotIndex
      properties:
        fromSlotIndex:
          type: integer
          minimum: 0
          maximum: 99
        toSlotIndex:
          type: integer
          minimum: 0
          maximum: 99
    SortInventoryRequest:
      type: object
      required:
        - sortBy
      properties:
        sortBy:
          $ref: "#/components/schemas/InventorySortOption"
...
    LootInventoryRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: object
            required:
              - itemId
              - quantity
            properties:
              itemId:
                type: string
                pattern: "^\\d+$"
              quantity:
                type: integer
                minimum: 1
                maximum: 9999
...
    Character:
      type: object
      required:
        - id
        - slotIndex
        - name
        - gender
        - job
        - progressionRank
        - location
        - level
        - exp
        - penya
        - stats
        - skillLevels
        - consumableLoadout
        - equipment
        - equipmentSets
        - ammoQuantity
        - ammoQuantities
        - activeQuestIds
        - completedQuestIds
        - inventory
      properties:
        id:
          type: string
        slotIndex:
          type: integer
          minimum: 0
          maximum: 7
        name:
          type: string
        gender:
          $ref: "#/components/schemas/CharacterGender"
        job:
          type: string
        progressionRank:
          $ref: "#/components/schemas/CharacterProgressionRank"
        location:
          type: string
        level:
          type: integer
          minimum: 1
        exp:
          type: integer
          minimum: 0
        penya:
          type: integer
          minimum: 0
        stats:
          $ref: "#/components/schemas/CharacterStats"
        skillLevels:
          $ref: "#/components/schemas/CharacterSkillLevels"
        consumableLoadout:
          $ref: "#/components/schemas/CharacterConsumableLoadout"
        equipment:
          $ref: "#/components/schemas/CharacterEquipment"
        equipmentSets:
          type: array
          minItems: 3
          maxItems: 3
          items:
            $ref: "#/components/schemas/CharacterEquipment"
        ammoQuantity:
          type: integer
          minimum: 0
        ammoQuantities:
          type: array
          minItems: 3
          maxItems: 3
          items:
            type: integer
            minimum: 0
        activeQuestIds:
          type: array
          items:
            type: integer
            minimum: 1
        completedQuestIds:
          type: array
          items:
            type: integer
            minimum: 1
        inventory:
          $ref: "#/components/schemas/CharacterInventory"
    CharacterGender:
      type: string
      enum:
        - male
        - female
    CharacterProgressionRank:
      type: string
      enum:
        - normal
        - master
        - hero
    CharacterStats:
      type: object
      required:
        - str
        - sta
        - dex
        - int
      properties:
        str:
          type: integer
          minimum: 15
          maximum: 999
          example: 15
        sta:
          type: integer
          minimum: 15
          maximum: 999
          example: 15
        dex:
          type: integer
          minimum: 15
          maximum: 999
          example: 15
        int:
          type: integer
          minimum: 15
          maximum: 999
          example: 15
    CharacterSkillLevels:
      type: object
      additionalProperties:
        type: integer
        minimum: 0
        maximum: 20
      example:
        vagrant-clean-hit: 1
...
    CharacterConsumableSlot:
      type: object
      nullable: true
      required:
        - itemId
        - quantity
      properties:
        itemId:
          type: string
        quantity:
          type: integer
          minimum: 1
    CharacterConsumableLoadout:
      type: object
      required:
        - hp
        - mp
        - fp
      properties:
        hp:
          $ref: "#/components/schemas/CharacterConsumableSlot"
        mp:
          $ref: "#/components/schemas/CharacterConsumableSlot"
        fp:
          $ref: "#/components/schemas/CharacterConsumableSlot"
    CharacterEquipment:
      type: object
      required:
        - helmet
        - suit
        - gloves
        - boots
        - flying
        - csBoots
        - csGloves
        - csSuit
        - csHelm
        - mask
        - cloak
        - ammo
        - offhand
        - mainhand
        - ringR
        - earringR
        - necklace
        - earringL
        - ringL
      properties:
        helmet:
          type: string
          nullable: true
        suit:
          type: string
          nullable: true
        gloves:
          type: string
          nullable: true
        boots:
          type: string
          nullable: true
        flying:
          type: string
          nullable: true
        csBoots:
          type: string
          nullable: true
        csGloves:
          type: string
          nullable: true
        csSuit:
          type: string
          nullable: true
        csHelm:
          type: string
          nullable: true
        mask:
          type: string
          nullable: true
        cloak:
          type: string
          nullable: true
        ammo:
          type: string
          nullable: true
        offhand:
          type: string
          nullable: true
        mainhand:
          type: string
          nullable: true
        ringR:
          type: string
          nullable: true
        earringR:
          type: string
          nullable: true
        necklace:
          type: string
          nullable: true
        earringL:
          type: string
          nullable: true
        ringL:
          type: string
          nullable: true
...
    CharacterInventory:
      type: object
      required:
        - size
        - items
      properties:
        size:
          type: integer
          minimum: 1
          example: 50
        items:
          type: array
          items:
            $ref: "#/components/schemas/CharacterInventoryItem"
    CharacterInventoryItem:
      type: object
      required:
        - slotIndex
        - itemId
        - quantity
      properties:
        slotIndex:
          type: integer
          minimum: 0
        itemId:
          type: string
        quantity:
          type: integer
          minimum: 1
...
    InventorySortOption:
      type: string
      enum:
        - name
        - level
        - job
        - category
...
    ErrorResponse:
      type: object
      required:
        - code
        - error
      properties:
        code:
          type: string
          enum:
            - authentication_required
            - conflict
            - domain_rule_failed
            - forbidden
            - internal_error
            - invalid_request
            - not_found
            - upstream_error
        error:
          type: string